+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.1.5] (2026-03-18)
+
+### Bug Fixes
+
+- **[LOCALE]** Fixed missing key for Charges in OperationsView
+
+
+### Features
+
+- [UI] Changed Buy/Sell Radio buttons to a Switch
+- [UI] Added ETF type selection Radio buttons in the ETF tab, to enable future support of Money Market ETFs and Bonds ETFs
+
+
+
+
+
 ## [0.1.4] (2026-03-18)
 
 ### Bug Fixes